Keep the behind_threshold columns so upgrading does not change this schema (#5173)
#5184 removed warning_behind_threshold / critical_behind_threshold from three places: the
DDL, the extended-tracking SELECT, and ShardStateSelector's ordinals. The reasoning holds
-- nothing in any repo ever wrote or read them, and every lag threshold lives in the
monitoring tool -- and the read-side half of it stands. Only the DDL is restored.

The cost of removing them was never the two always-NULL columns, it is the schema CHANGE.
Measured against PostgreSQL, on a store with event sourcing registered, an
mt_event_progression carrying a column the configuration no longer declares gives:

    CreateMigrationAsync().Difference   = Update   (mt_event_progression=Update)
    AssertDatabaseMatchesConfigurationAsync()      -> throws
    ApplyAllConfiguredChangesToDatabaseAsync()     -> drops the column

So an upgrade is not the no-op #5184's commit message implied. Two things happen to
existing deployments. A host running AutoCreate.None with a schema assertion at boot --
the standard production posture -- does not migrate, it refuses to start until someone
migrates it. And a host that does apply runs `alter table ... drop column`, which in
PostgreSQL is O(1) metadata but needs ACCESS EXCLUSIVE on a small, hot table that every
running daemon writes; while that lock waits it blocks every reader and writer behind it.

Two unused columns are much cheaper than asking every deployment to absorb that, and
keeping them also means a freshly created database and an upgraded one have the same
shape rather than diverging on a detail nothing reads.

If they are ever removed for real it has to be a documented migration step rather than a
silent consequence of upgrading a package.

The tests pin both halves and both fail with the columns taken back out. Note the
`AddEventType` in each: without an event type registered the event store's tables are not
part of the schema migration at all, every delta comes back None, and any assertion about
this table passes vacuously -- which is how the original version of this test was green
against both shapes.
