+++ src/Marten/Events/Schema/EventProgressionTable.cs
@@ -51,13 +51,25 @@ internal class EventProgressionTable: Table
             AddColumn("pause_reason", "text").AllowNulls();
             AddColumn("running_on_node", "integer").AllowNulls();
 
-            // #5173: warning_behind_threshold / critical_behind_threshold used to be created here.
-            // They were created, selected and hydrated onto ShardState -- and written by nothing, in
-            // any repo, so every read of them returned NULL. Two columns of storage, two entries in
-            // every extended-tracking SELECT and two selector ordinals, carrying nothing. Removed
-            // rather than wired: nothing ever owned the value. Existing deployments get an
-            // `alter table ... drop column` on their next apply, which is lossless because the
-            // columns were provably always NULL.
+            // #5173: warning_behind_threshold / critical_behind_threshold are DELIBERATELY still
+            // created, and are the one part of the extended block that is neither written nor read.
+            // Nothing in any repo ever owned the value, so #5184 removed them outright -- from the
+            // DDL, the SELECT, and the ShardStateSelector ordinals. The read-side half of that stands;
+            // only the DDL is restored here.
+            //
+            // Restored because the cost of removing them is not the two columns, it is the schema
+            // CHANGE. Dropping a column from an existing deployment means `alter table ... drop
+            // column` on the next apply, and in PostgreSQL that needs ACCESS EXCLUSIVE on a small,
+            // hot table that every running daemon writes. The operation itself is O(1) metadata, but
+            // the lock queues behind in-flight progression writes and blocks every reader and writer
+            // behind it while it waits. Two always-NULL columns are cheaper than asking every
+            // deployment to take that lock, so an upgrade stays a no-op on this table.
+            //
+            // They are not selected and not hydrated -- see ProjectionProgressStatement and
+            // ShardStateSelector, whose ordinals now skip them. If they are ever removed for real,
+            // it has to be a documented migration step, not a silent auto-apply.
+            AddColumn("warning_behind_threshold", "bigint").AllowNulls();
+            AddColumn("critical_behind_threshold", "bigint").AllowNulls();
 
             // #5048 / jasperfx#565: the classified reason this shard is paused or stopped, so a consumer
             // polling the database (CritterWatch when the publishing node is DOWN, which is exactly when
